Fix integration test mongoDBOnline checks
Upped logging - so can debug jenkins

+++ driver/src/test/scala/org/mongodb/scala/helpers/RequiresMongoDBSpec.scala
@@ -96,7 +96,7 @@ trait RequiresMongoDBSpec extends FlatSpec with Matchers with ScalaFutures with 
   }
 
   override def afterAll() {
-    Await.result(mongoClient(databaseName).admin.drop(), WAIT_DURATION)
+    if (mongoDbOnline) Await.result(mongoClient(databaseName).admin.drop(), WAIT_DURATION)
   }
 
 }
